Update README.md
I changed the size of the bubble by going to line 9 and changing the radius from 50 to 100. Changed the shake by going to lines 27 and 28 and changing the (-1, 1) values to (-5, 5) respectively. For the color, I went  to line 23 and changed one of the fill values of 255 to 50. I was able to do the challenge. I simply got rid of lines 15 and 19 and changed the x and y to equals mouseX and mouseY respectively instead of having random(width) and random(height).